+++ Hurtownia/Windows/DeliveryWindow.xaml.cs
@@ -31,13 +31,13 @@ namespace Hurtownia.Windows
                 var name = ComboBoxProducts.Text;
 
                 var product = Products.GetProduct(name);
-                product.Price = float.Parse(TextBoxPrice.Text);
-                product.Quantity = float.Parse(TextBoxQuantity.Text);
+                product.Price = double.Parse(TextBoxPrice.Text);
+                product.Quantity = double.Parse(TextBoxQuantity.Text);
                 product.Cost = product.Price*product.Quantity;
 
                 ////Product = Products.GetProduct(name);
-                //Product.Price = float.Parse(TextBoxPrice.Text);
-                //Product.Quantity = float.Parse(TextBoxQuantity.Text);
+                //Product.Price = double.Parse(TextBoxPrice.Text);
+                //Product.Quantity = double.Parse(TextBoxQuantity.Text);
                 //Product.Cost = Product.Price*Product.Quantity;
 
                 newDelivery.AddProduct(product);
